fix build process
